+++ suager/cogs/tags.py
@@ -96,7 +96,7 @@ class Tags(commands.Cog):
                 # content = "Okay then, nothing will be deleted..."
                 return await confirm_msg.edit(content=langs.gls("generic_timed_out", locale, confirm_msg.clean_content))
             self.db.execute("DELETE FROM tags WHERE gid=? AND name=?", (ctx.guild.id, tag['name']))
-            return await general.send(langs.gls("tags_delete_success", locale), ctx.channel)
+            return await general.send(langs.gls("tags_delete_success", locale, tag["name"]), ctx.channel)
             # return await general.send(f"The tag `{tag['name']}` has been successfully deleted.", ctx.channel)
         else:
             return await general.send(langs.gls("tags_delete_deny", locale), ctx.channel)