+++ ChangeLog.md
@@ -205,3 +205,8 @@
 ## v6.0-beta3 - 4 August
 - Support for floats in base conversions from base 10
 - Removed owoify and neon commands due to no usage
+
+## v6.0-beta4 - 5 August
+- Removed global levels
+- Improved levels
+- Fixed some broken outputs